requirements.txt
clu
flax
imageio
jax
lpips
matplotlib
mmcv==1.7.1
yapf==0.40.0
networkx
numpy
Pillow
git+https://github.com/lucasb-eyer/pydensecrf.git
pytorch_msssim
scipy
scikit_image
scikit_learn
timm
torch_efficient_distloss
tqdm
Ninja
einops
imageio-ffmpeg
tensorboard
ftfy
regex